Move all plugin related methods to button, implement __getattr__ as plot proxy
